refactor: modularize platform-specific logic
