+++ e2e-app/app-config-tools/arch.ts
@@ -1,0 +1,9 @@
+let newArch = false
+
+export function disableNewArch() {
+  newArch = false
+}
+
+export function getNewArch() {
+  return newArch
+}
